backend/requirements/_base.txt
asgiref==3.5.2
celery==5.2.7
Django==4.0.5
django-admin-autocomplete-filter==0.7.1
django-anymail[sendgrid]==8.6
django-celery-results==2.4.0
django-ckeditor==6.4.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-polymorphic==3.1.0
django-redis==5.2.0
django-rest-polymorphic==0.1.9
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
gunicorn==20.1.0
mongoengine==0.24.0
mysqlclient==2.1.0
openpyxl==3.0.10
Pillow==9.1.1
PyAMQP==0.1.0.7
qrcode[pil]==7.3.1
redis==4.3.3
singlemodeladmin==0.9
sqlparse==0.4.2
transbank-sdk==2.0.1
uvicorn[standard]==0.17.6
weasyprint==55.0
rut_chile==2.0.1
